// services/authService/src/interfaces/http/common/errors/problem-code.ts
export enum ProblemCode {
    BadRequest = 'BAD_REQUEST',
    Unauthorized = 'UNAUTHORIZED',
    Forbidden = 'FORBIDDEN',
    NotFound = 'NOT_FOUND',
    MethodNotAllowed = 'METHOD_NOT_ALLOWED',
    Conflict = 'CONFLICT',
    PayloadTooLarge = 'PAYLOAD_TOO_LARGE',
    UnsupportedMediaType = 'UNSUPPORTED_MEDIA_TYPE',
    UnprocessableEntity = 'UNPROCESSABLE_ENTITY',
    ValidationError = 'VALIDATION_ERROR',
    TooManyRequests = 'TOO_MANY_REQUESTS',
    InternalServerError = 'INTERNAL_SERVER_ERROR',
    BadGateway = 'BAD_GATEWAY',
    ServiceUnavailable = 'SERVICE_UNAVAILABLE',
    GatewayTimeout = 'GATEWAY_TIMEOUT',
    HttpError = 'HTTP_ERROR',
}